Extract AOC capability core module and isolate protocol layer
