+++ aoc/capabilities/core/index.ts
@@ -1,0 +1,13 @@
+export { evaluateCapabilityAccess } from './evaluateCapabilityAccess';
+export { capabilityAccessReasonCodes } from './reasonCodes';
+export type { CapabilityAccessReasonCode } from './reasonCodes';
+export type {
+  CapabilityAccessDecision,
+  CapabilityAccessHookResult,
+  CapabilityAccessRequest,
+  CapabilityAccessChecks,
+  CapabilityResource,
+  NormalizedCapabilityAccessRequest,
+  CapabilityPolicyHook,
+  CapabilityUsageHook
+} from './types';
