+++ aoc/capabilities/index.ts
@@ -1,0 +1,36 @@
+export { evaluateCapabilityAccess } from './core/evaluateCapabilityAccess';
+export { consumeCapabilityAccess } from './runtime/consumeCapabilityAccess';
+export { interpretWithCapability } from './interpreter/interpretWithCapability';
+
+export {
+  mintCapabilityToken,
+  validateCapabilityToken,
+  verifyCapabilityToken,
+  resetNonceRegistry
+} from './tokens/capabilityToken';
+
+export { MarketMakerRegistry } from './market/marketMakerRegistry';
+
+export { capabilityAccessReasonCodes } from './core/reasonCodes';
+export { capabilityConsumptionReasonCodes, InMemoryConsentUsageRegistry } from './runtime/consumeCapabilityAccess';
+export { enforceCapability } from './legacy/capabilityEnforcer';
+
+export type {
+  CapabilityAccessDecision,
+  CapabilityAccessRequest,
+  CapabilityAccessReasonCode,
+  CapabilityAccessChecks,
+  CapabilityPolicyHook,
+  CapabilityUsageHook,
+  CapabilityResource
+} from './core';
+
+export type {
+  CapabilityConsumptionChecks,
+  CapabilityConsumptionDecision,
+  CapabilityConsumptionReasonCode,
+  CapabilityConsumptionRequest,
+  ConsentUsageRegistry,
+  ConsentUsageState,
+  ConsentUsageResult
+} from './runtime';
