+++ aoc/capabilities/interpreter/index.ts
@@ -1,0 +1,1 @@
+export { interpretWithCapability, runInterpreter } from './interpretWithCapability';
