+++ aoc/capabilities/interpreter/interpretWithCapability.ts
@@ -1,0 +1,1 @@
+export { interpretWithCapability, runInterpreter } from '../../../interpreter/aiInterpreter';
