+++ aoc/capabilities/legacy/index.ts
@@ -1,0 +1,2 @@
+export { enforceCapability } from './capabilityEnforcer';
+export type { EnforceCapabilityDecision, EnforceCapabilityInput } from './capabilityEnforcer';
