+++ aoc/capabilities/market/marketMakerRegistry.ts
@@ -1,0 +1,5 @@
+export {
+  MarketMakerRegistry,
+  type MarketMaker,
+  type MarketMakerStatus
+} from '../../../shared/marketMakerRegistry';
