+++ aoc/capabilities/tokens/canonical.ts
@@ -1,0 +1,2 @@
+export { canonicalizeCapabilityPayload } from '../../../capability/canonical';
+export type { CapabilityPayload } from '../../../capability/canonical';
