+++ aoc/capabilities/tokens/capabilityToken.ts
@@ -1,0 +1,7 @@
+export {
+  mintCapabilityToken,
+  validateCapabilityToken,
+  verifyCapabilityToken,
+  resetNonceRegistry
+} from '../../../capability/capabilityToken';
+export type { CapabilityTokenV1, MintCapabilityOptions, ScopeEntry } from '../../../capability/types';
