+++ aoc/capabilities/tokens/hash.ts
@@ -1,0 +1,1 @@
+export { computeCapabilityHash } from '../../../capability/hash';
