+++ enforcement/reasonCodes.ts
@@ -1,28 +1,4 @@
-export const capabilityAccessReasonCodes = {
-  ACCESS_ALLOWED: 'ACCESS_ALLOWED',
-  CAPABILITY_MISSING: 'CAPABILITY_MISSING',
-  CAPABILITY_INVALID: 'CAPABILITY_INVALID',
-  CAPABILITY_INACTIVE: 'CAPABILITY_INACTIVE',
-  CAPABILITY_NOT_YET_VALID: 'CAPABILITY_NOT_YET_VALID',
-  CAPABILITY_EXPIRED: 'CAPABILITY_EXPIRED',
-  CONSENT_EXPIRED: 'CONSENT_EXPIRED',
-  EXPIRATION_MISMATCH: 'EXPIRATION_MISMATCH',
-  ACTION_MISSING: 'ACTION_MISSING',
-  ACTION_NOT_ALLOWED: 'ACTION_NOT_ALLOWED',
-  RESOURCE_MISSING: 'RESOURCE_MISSING',
-  RESOURCE_NOT_ALLOWED: 'RESOURCE_NOT_ALLOWED',
-  MARKET_MAKER_REQUIRED: 'MARKET_MAKER_REQUIRED',
-  MARKET_MAKER_MISMATCH: 'MARKET_MAKER_MISMATCH',
-  UNKNOWN_MARKET_MAKER: 'UNKNOWN_MARKET_MAKER',
-  MARKET_MAKER_DEPRECATED: 'MARKET_MAKER_DEPRECATED',
-  MARKET_MAKER_REVOKED: 'MARKET_MAKER_REVOKED',
-  POLICY_DENIED: 'POLICY_DENIED',
-  USAGE_DENIED: 'USAGE_DENIED',
-  PAYMENT_REQUIRED: 'PAYMENT_REQUIRED',
-  CONSENT_INVALID: 'CONSENT_INVALID',
-  CONSENT_MISMATCH: 'CONSENT_MISMATCH',
-  INTERNAL_EVALUATION_ERROR: 'INTERNAL_EVALUATION_ERROR'
-} as const;
-
-export type CapabilityAccessReasonCode =
-  (typeof capabilityAccessReasonCodes)[keyof typeof capabilityAccessReasonCodes];
+export {
+  capabilityAccessReasonCodes,
+  type CapabilityAccessReasonCode
+} from '../aoc/capabilities/core/reasonCodes';
