+++ index.ts
@@ -7,6 +7,7 @@ export * from './field';
 export * from './pack';
 export * from './storage';
 export * from './protocol/capabilities';
+export * from './aoc/capabilities';
 export * from './interpreter';
 
 export * from './enforcement';
